quic: add code documentation + minor refactors
