+++ lib/internal/quic/core.js
@@ -18,6 +18,8 @@ const {
   getSocketType,
   lookup4,
   lookup6,
+  setTransportParams,
+  toggleListeners,
   validateCloseCode,
   validateTransportParams,
   validateQuicClientSessionOptions,
@@ -110,8 +112,6 @@ const {
   createClientSession: _createClientSession,
   openBidirectionalStream: _openBidirectionalStream,
   openUnidirectionalStream: _openUnidirectionalStream,
-  sessionConfig,
-  http3Config,
   setCallbacks,
   constants: {
     AF_INET,
@@ -119,28 +119,7 @@ const {
     NGTCP2_ALPN_H3,
     NGTCP2_MAX_CIDLEN,
     NGTCP2_MIN_CIDLEN,
-    IDX_QUIC_SESSION_ACTIVE_CONNECTION_ID_LIMIT,
-    IDX_QUIC_SESSION_MAX_STREAM_DATA_BIDI_LOCAL,
-    IDX_QUIC_SESSION_MAX_STREAM_DATA_BIDI_REMOTE,
-    IDX_QUIC_SESSION_MAX_STREAM_DATA_UNI,
-    IDX_QUIC_SESSION_MAX_DATA,
-    IDX_QUIC_SESSION_MAX_STREAMS_BIDI,
-    IDX_QUIC_SESSION_MAX_STREAMS_UNI,
-    IDX_QUIC_SESSION_IDLE_TIMEOUT,
-    IDX_QUIC_SESSION_MAX_PACKET_SIZE,
-    IDX_QUIC_SESSION_MAX_CRYPTO_BUFFER,
-    IDX_QUIC_SESSION_CONFIG_COUNT,
-    IDX_HTTP3_QPACK_MAX_TABLE_CAPACITY,
-    IDX_HTTP3_QPACK_BLOCKED_STREAMS,
-    IDX_HTTP3_MAX_HEADER_LIST_SIZE,
-    IDX_HTTP3_MAX_PUSHES,
-    IDX_HTTP3_CONFIG_COUNT,
     IDX_QUIC_SESSION_MAX_PACKET_SIZE_DEFAULT,
-    IDX_QUIC_SESSION_MAX_ACK_DELAY,
-    IDX_QUIC_SESSION_STATE_CERT_ENABLED,
-    IDX_QUIC_SESSION_STATE_CLIENT_HELLO_ENABLED,
-    IDX_QUIC_SESSION_STATE_PATH_VALIDATED_ENABLED,
-    IDX_QUIC_SESSION_STATE_KEYLOG_ENABLED,
     IDX_QUIC_SESSION_STATE_MAX_STREAMS_BIDI,
     IDX_QUIC_SESSION_STATE_MAX_STREAMS_UNI,
     IDX_QUIC_SESSION_STATE_MAX_DATA_LEFT,
@@ -244,126 +223,71 @@ const kSocketDestroyed = 4;
 
 let diagnosticPacketLossWarned = false;
 
-function setConfigField(buffer, val, index) {
-  if (typeof val === 'number') {
-    buffer[index] = val;
-    return 1 << index;
-  }
-  return 0;
-}
-
-function setTransportParams(config) {
-  const {
-    activeConnectionIdLimit,
-    maxStreamDataBidiLocal,
-    maxStreamDataBidiRemote,
-    maxStreamDataUni,
-    maxData,
-    maxStreamsBidi,
-    maxStreamsUni,
-    idleTimeout,
-    maxPacketSize,
-    maxAckDelay,
-    maxCryptoBuffer,
-    h3: {
-      qpackMaxTableCapacity,
-      qpackBlockedStreams,
-      maxHeaderListSize,
-      maxPushes,
-    },
-  } = { h3: {}, ...config };
-
-  const flags = setConfigField(sessionConfig,
-                               activeConnectionIdLimit,
-                               IDX_QUIC_SESSION_ACTIVE_CONNECTION_ID_LIMIT) |
-                setConfigField(sessionConfig,
-                               maxStreamDataBidiLocal,
-                               IDX_QUIC_SESSION_MAX_STREAM_DATA_BIDI_LOCAL) |
-                setConfigField(sessionConfig,
-                               maxStreamDataBidiRemote,
-                               IDX_QUIC_SESSION_MAX_STREAM_DATA_BIDI_REMOTE) |
-                setConfigField(sessionConfig,
-                               maxStreamDataUni,
-                               IDX_QUIC_SESSION_MAX_STREAM_DATA_UNI) |
-                setConfigField(sessionConfig,
-                               maxData,
-                               IDX_QUIC_SESSION_MAX_DATA) |
-                setConfigField(sessionConfig,
-                               maxStreamsBidi,
-                               IDX_QUIC_SESSION_MAX_STREAMS_BIDI) |
-                setConfigField(sessionConfig,
-                               maxStreamsUni,
-                               IDX_QUIC_SESSION_MAX_STREAMS_UNI) |
-                setConfigField(sessionConfig,
-                               idleTimeout * 1000000,  // millis to nano
-                               IDX_QUIC_SESSION_IDLE_TIMEOUT) |
-                setConfigField(sessionConfig,
-                               maxAckDelay,
-                               IDX_QUIC_SESSION_MAX_ACK_DELAY) |
-                setConfigField(sessionConfig,
-                               maxPacketSize,
-                               IDX_QUIC_SESSION_MAX_PACKET_SIZE) |
-                setConfigField(sessionConfig,
-                               maxCryptoBuffer,
-                               IDX_QUIC_SESSION_MAX_CRYPTO_BUFFER);
-
-  sessionConfig[IDX_QUIC_SESSION_CONFIG_COUNT] = flags;
-
-  const h3flags = setConfigField(http3Config,
-                                 qpackMaxTableCapacity,
-                                 IDX_HTTP3_QPACK_MAX_TABLE_CAPACITY) |
-                  setConfigField(http3Config,
-                                 qpackBlockedStreams,
-                                 IDX_HTTP3_QPACK_BLOCKED_STREAMS) |
-                  setConfigField(http3Config,
-                                 maxHeaderListSize,
-                                 IDX_HTTP3_MAX_HEADER_LIST_SIZE) |
-                  setConfigField(http3Config,
-                                 maxPushes,
-                                 IDX_HTTP3_MAX_PUSHES);
-
-  http3Config[IDX_HTTP3_CONFIG_COUNT] = h3flags;
-}
-
-// Called when the socket is closed
+// Called by the C++ internals when the socket is closed.
+// When this is called, the only thing left to do is destroy
+// the QuicSocket instance.
 function onSocketClose() {
   this[owner_symbol].destroy();
 }
 
-// Called when an error occurs on the socket
+// Called by the C++ internals when an error occurs on the socket.
+// When this is called, the only thing left to do is destroy
+// the QuicSocket instance with an error.
+// TODO(@jasnell): Should consolidate this with onSocketClose
 function onSocketError(err) {
   this[owner_symbol].destroy(errnoException(err));
 }
 
+// Called by the C++ internals when the server busy state of
+// the QuicSocket has been changed.
 function onSocketServerBusy(on) {
   this[owner_symbol][kServerBusy](!!on);
 }
 
-// Called when a new QuicSession is ready to use
-function onSessionReady(sessionHandle) {
+// Called by the C++ internals when a new server QuicSession has been created.
+function onSessionReady(handle) {
   const socket = this[owner_symbol];
-  const session = new QuicServerSession(socket, sessionHandle);
+  const session = new QuicServerSession(socket, handle);
   process.nextTick(emit.bind(socket, 'session', session));
 }
 
+// During an immediate close, all currently open QuicStreams are
+// abruptly closed. If they are still writable or readable, an abort
+// event will be emitted, and RESET_STREAM and STOP_SENDING frames
+// will be transmitted as necessary. Once streams have been
+// shutdown, a CONNECTION_CLOSE frame will be sent and the
+// session will enter the closing period, after which it will
+// be destroyed either when the idle timeout expires, the
+// QuicSession is silently closed, or destroy is called.
 function onSessionClose(code, family) {
-  // During an immediate close, all currently open QuicStreams are
-  // abruptly closed. If they are still writable or readable, an abort
-  // event will be emitted, and RESET_STREAM and STOP_SENDING frames
-  // will be transmitted as necessary. Once streams have been
-  // shutdown, a CONNECTION_CLOSE frame will be sent and the
-  // session will enter the closing period, after which it will
-  // be destroyed either when the idle timeout expires, the
-  // QuicSession is silently closed, or destroy is called.
   this[owner_symbol][kClose](family, code);
 }
 
-function onSessionDestroyed(code, family) {
+// Called by the C++ internals when a QuicSession has been destroyed.
+// When this is called, the QuicSession is no longer usable. Removing
+// the handle and emitting close is the only action.
+// TODO(@jasnell): In the future, this will need to act differently
+// for QuicClientSessions when autoResume is enabled.
+function onSessionDestroyed() {
   const session = this[owner_symbol];
   this[owner_symbol] = undefined;
 
   session[kSetHandle]();
   process.nextTick(emit.bind(session, 'close'));
+}
+
+// Used only within the onSessionClientHello function. Invoked
+// to complete the client hello process.
+function clientHelloCallback(err, ...args) {
+  if (err) {
+    this[owner_symbol].destroy(err);
+    return;
+  }
+  try {
+    this.onClientHelloDone(...args);
+  } catch (err) {
+    this[owner_symbol].destroy(err);
+  }
 }
 
 // This callback is invoked at the start of the TLS handshake to provide
@@ -375,17 +299,39 @@ function onSessionClientHello(alpn, servername, ciphers) {
     alpn,
     servername,
     ciphers,
-    (err, ...args) => {
-      if (err) {
-        this[owner_symbol].destroy(err);
-        return;
-      }
-      try {
-        this.onClientHelloDone(...args);
-      } catch (err) {
-        this[owner_symbol].destroy(err);
-      }
-    });
+    clientHelloCallback.bind(this));
+}
+
+// Used only within the onSessionCert function. Invoked
+// to complete the session cert process.
+function sessionCertCallback(err, context, ocspResponse) {
+  if (err) {
+    this[owner_symbol].destroy(err);
+    return;
+  }
+  if (context != null && !context.context) {
+    this[owner_symbol].destroy(
+      new ERR_INVALID_ARG_TYPE(
+        'context',
+        'SecureContext',
+        context));
+  }
+  if (ocspResponse != null) {
+    if (typeof ocspResponse === 'string')
+      ocspResponse = Buffer.from(ocspResponse);
+    if (!isArrayBufferView(ocspResponse)) {
+      this[owner_symbol].destroy(
+        new ERR_INVALID_ARG_TYPE(
+          'ocspResponse',
+          ['string', 'Buffer', 'TypedArray', 'DataView'],
+          ocspResponse));
+    }
+  }
+  try {
+    this.onCertDone(context ? context.context : undefined, ocspResponse);
+  } catch (err) {
+    this[owner_symbol].destroy(err);
+  }
 }
 
 // This callback is only ever invoked for QuicServerSession instances,
@@ -393,35 +339,7 @@ function onSessionClientHello(alpn, servername, ciphers) {
 // user callback must invoke .onCertDone() in order for the
 // TLS handshake to continue.
 function onSessionCert(servername) {
-  this[owner_symbol][kCert](servername, (err, context, ocspResponse) => {
-    if (err) {
-      this[owner_symbol].destroy(err);
-      return;
-    }
-    if (context != null && !context.context) {
-      this[owner_symbol].destroy(
-        new ERR_INVALID_ARG_TYPE(
-          'context',
-          'SecureContext',
-          context));
-    }
-    if (ocspResponse != null) {
-      if (typeof ocspResponse === 'string')
-        ocspResponse = Buffer.from(ocspResponse);
-      if (!isArrayBufferView(ocspResponse)) {
-        this[owner_symbol].destroy(
-          new ERR_INVALID_ARG_TYPE(
-            'ocspResponse',
-            ['string', 'Buffer', 'TypedArray', 'DataView'],
-            ocspResponse));
-      }
-    }
-    try {
-      this.onCertDone(context ? context.context : undefined, ocspResponse);
-    } catch (err) {
-      this[owner_symbol].destroy(err);
-    }
-  });
+  this[owner_symbol][kCert](servername, sessionCertCallback.bind(this));
 }
 
 // This callback is only ever invoked for QuicClientSession instances,
@@ -432,6 +350,7 @@ function onSessionStatus(response) {
   this[owner_symbol][kCert](response);
 }
 
+// Called by the C++ internals when the TLS handshake is completed.
 function onSessionHandshake(
   servername,
   alpn,
@@ -450,6 +369,10 @@ function onSessionHandshake(
     verifyErrorCode);
 }
 
+// Called by the C++ internals when TLS session ticket data is
+// available. This is generally most useful on the client side
+// where the session ticket needs to be persisted for session
+// resumption and 0RTT.
 function onSessionTicket(sessionID, sessionTicket, transportParams) {
   if (this[owner_symbol]) {
     process.nextTick(
@@ -462,6 +385,9 @@ function onSessionTicket(sessionID, sessionTicket, transportParams) {
   }
 }
 
+// Called by the C++ internals when path validation is completed.
+// This is a purely informational event that is emitted only when
+// there is a listener present for the pathValidation event.
 function onSessionPathValidation(res, local, remote) {
   const session = this[owner_symbol];
   if (session) {
@@ -475,6 +401,7 @@ function onSessionPathValidation(res, local, remote) {
   }
 }
 
+// Called by the C++ internals to emit a QLog record.
 function onSessionQlog(str) {
   if (this.qlogBuffer === undefined) this.qlogBuffer = '';
 
@@ -507,13 +434,16 @@ function onSessionQlog(str) {
   }
 }
 
-// Called when an error occurs in a QuicSession
+// Called when an error occurs in a QuicSession. When this happens,
+// the only remedy is to destroy the session.
 function onSessionError(error) {
   if (this[owner_symbol]) {
     this[owner_symbol].destroy(error);
   }
 }
 
+// Called by the C++ internals when a client QuicSession receives
+// a version negotiation response from the server.
 function onSessionVersionNegotiation(
   version,
   requestedVersions,
@@ -526,13 +456,15 @@ function onSessionVersionNegotiation(
   }
 }
 
+// Called by the C++ internals to emit keylogging details for a
+// QuicSession.
 function onSessionKeylog(line) {
   if (this[owner_symbol]) {
     this[owner_symbol].emit('keylog', line);
   }
 }
 
-// Called when a new QuicStream is ready to use
+// Called by the C++ internals when a new QuicStream has been created.
 function onStreamReady(streamHandle, id) {
   const session = this[owner_symbol];
 
@@ -551,12 +483,13 @@ function onStreamReady(streamHandle, id) {
   process.nextTick(emit.bind(session, 'stream', stream));
 }
 
-// Called when a stream is closed on the C++ side and
+// Called by the C++ internals when a stream is closed and
 // needs to be destroyed on the JavaScript side.
 function onStreamClose(id, appErrorCode) {
   this[owner_symbol][kStreamClose](id, appErrorCode);
 }
 
+// Called by the C++ internals when a stream has been reset
 function onStreamReset(id, appErrorCode, finalSize) {
   this[owner_symbol][kStreamReset](id, appErrorCode, finalSize);
 }
@@ -574,11 +507,11 @@ function onStreamHeaders(id, headers, kind) {
   this[owner_symbol][kHeaders](id, headers, kind);
 }
 
+// During a silent close, all currently open QuicStreams are abruptly
+// closed. If they are still writable or readable, an abort event will be
+// emitted, otherwise the stream is just destroyed. No RESET_STREAM or
+// STOP_SENDING is transmitted to the peer.
 function onSessionSilentClose(statelessReset, code, family) {
-  // During a silent close, all currently open QuicStreams are abruptly
-  // closed. If they are still writable or readable, an abort event will be
-  // emitted, otherwise the stream is just destroyed. No RESET_STREAM or
-  // STOP_SENDING is transmitted to the peer.
   this[owner_symbol][kDestroy](statelessReset, family, code);
 }
 
@@ -608,6 +541,11 @@ setCallbacks({
   onStreamHeaders,
 });
 
+// afterLookup is invoked when binding a QuicEndpoint. The first
+// step to binding is to resolve the given hostname into an ip
+// address. Once resolution is complete, the ip address needs to
+// be passed on to the [kContinueBind] function or the QuicEndpoint
+// needs to be destroyed.
 function afterLookup(err, ip) {
   if (err) {
     this.destroy(err);
@@ -616,6 +554,11 @@ function afterLookup(err, ip) {
   this[kContinueBind](ip);
 }
 
+// connectAfterLookup is invoked when the QuicSocket connect()
+// method has been invoked. The first step is to resolve the given
+// remote hostname into an ip address. Once resolution is complete,
+// the resolved ip address needs to be passed on to the [kContinueConnect]
+// function or the QuicClientSession needs to be destroyed.
 function connectAfterLookup(type, err, ip) {
   if (err) {
     this.destroy(err);
@@ -624,6 +567,10 @@ function connectAfterLookup(type, err, ip) {
   this[kContinueConnect](type, ip);
 }
 
+// Used only from within the continueListen function. When a preferred address
+// has been provided, the hostname given must be resolved into an IP address,
+// which must be passed on to [kContinueListen] or the QuicSocket needs to be
+// destroyed.
 function afterPreferredAddressLookup(
   transportParams,
   port,
@@ -637,6 +584,14 @@ function afterPreferredAddressLookup(
   this[kContinueListen](transportParams, { address, port, type });
 }
 
+// When the QuicSocket listen() function is called, the first step is to bind
+// the underlying QuicEndpoint's. Once at least one endpoint has been bound,
+// the continueListen function is invoked to continue the process of starting
+// to listen.
+//
+// The preferredAddress is a preferred network endpoint that the server wishes
+// connecting clients to use. If specified, this will be communicate to the
+// client as part of the tranport parameters exchanged during the TLS handshake.
 function continueListen(transportParams, lookup) {
   const { preferredAddress } = transportParams;
 
@@ -660,12 +615,19 @@ function continueListen(transportParams, lookup) {
   this[kContinueListen](transportParams);
 }
 
+// When the QuicSocket connect() function is called, the first step is to bind
+// the underlying QuicEndpoint's. Once at least one endpoint has been bound,
+// the connectAfterBind function is invoked to continue the connection process.
+//
+// The immediate next step is to resolve the address into an ip address.
 function connectAfterBind(session, lookup, address, type) {
   lookup(
     address || (type === AF_INET6 ? '::' : '0.0.0.0'),
     connectAfterLookup.bind(session, type));
 }
 
+// Creates the SecureContext used by QuicSocket instances that are listening
+// for new connections.
 function createSecureContext(options, init_cb) {
   const {
     ca,
@@ -710,43 +672,11 @@ function createSecureContext(options, init_cb) {
 }
 
 function onNewListener(event) {
-  if (this[kHandle] === undefined || this.listenerCount(event) !== 0)
-    return;
-
-  switch (event) {
-    case 'keylog':
-      this[kHandle].state[IDX_QUIC_SESSION_STATE_KEYLOG_ENABLED] = 1;
-      break;
-    case 'clientHello':
-      this[kHandle].state[IDX_QUIC_SESSION_STATE_CLIENT_HELLO_ENABLED] = 1;
-      break;
-    case 'pathValidation':
-      this[kHandle].state[IDX_QUIC_SESSION_STATE_PATH_VALIDATED_ENABLED] = 1;
-      break;
-    case 'OCSPRequest':
-      this[kHandle].state[IDX_QUIC_SESSION_STATE_CERT_ENABLED] = 1;
-      break;
-  }
+  toggleListeners(this, event, 1);
 }
 
 function onRemoveListener(event) {
-  if (this[kHandle] === undefined || this.listenerCount(event) !== 0)
-    return;
-
-  switch (event) {
-    case 'keylog':
-      this[kHandle].state[IDX_QUIC_SESSION_STATE_KEYLOG_ENABLED] = 0;
-      break;
-    case 'clientHello':
-      this[kHandle].state[IDX_QUIC_SESSION_STATE_CLIENT_HELLO_ENABLED] = 0;
-      break;
-    case 'pathValidation':
-      this[kHandle].state[IDX_QUIC_SESSION_STATE_PATH_VALIDATED_ENABLED] = 0;
-      break;
-    case 'OCSPRequest':
-      this[kHandle].state[IDX_QUIC_SESSION_STATE_CERT_ENABLED] = 0;
-      break;
-  }
+  toggleListeners(this, event, 0);
 }
 
 // QuicEndpoint wraps a UDP socket.
