+++ lib/internal/quic/util.js
@@ -19,12 +19,37 @@ const {
   isIP,
 } = require('internal/net');
 
+const { kHandle } = require('internal/stream_base_commons');
+
 const {
+  sessionConfig,
+  http3Config,
   constants: {
     AF_INET,
     AF_INET6,
     DEFAULT_RETRYTOKEN_EXPIRATION,
     DEFAULT_MAX_CONNECTIONS_PER_HOST,
+    IDX_QUIC_SESSION_ACTIVE_CONNECTION_ID_LIMIT,
+    IDX_QUIC_SESSION_MAX_STREAM_DATA_BIDI_LOCAL,
+    IDX_QUIC_SESSION_MAX_STREAM_DATA_BIDI_REMOTE,
+    IDX_QUIC_SESSION_MAX_STREAM_DATA_UNI,
+    IDX_QUIC_SESSION_MAX_DATA,
+    IDX_QUIC_SESSION_MAX_STREAMS_BIDI,
+    IDX_QUIC_SESSION_MAX_STREAMS_UNI,
+    IDX_QUIC_SESSION_IDLE_TIMEOUT,
+    IDX_QUIC_SESSION_MAX_ACK_DELAY,
+    IDX_QUIC_SESSION_MAX_PACKET_SIZE,
+    IDX_QUIC_SESSION_MAX_CRYPTO_BUFFER,
+    IDX_QUIC_SESSION_CONFIG_COUNT,
+    IDX_QUIC_SESSION_STATE_CERT_ENABLED,
+    IDX_QUIC_SESSION_STATE_CLIENT_HELLO_ENABLED,
+    IDX_QUIC_SESSION_STATE_KEYLOG_ENABLED,
+    IDX_QUIC_SESSION_STATE_PATH_VALIDATED_ENABLED,
+    IDX_HTTP3_QPACK_MAX_TABLE_CAPACITY,
+    IDX_HTTP3_QPACK_BLOCKED_STREAMS,
+    IDX_HTTP3_MAX_HEADER_LIST_SIZE,
+    IDX_HTTP3_MAX_PUSHES,
+    IDX_HTTP3_CONFIG_COUNT,
     MAX_RETRYTOKEN_EXPIRATION,
     MIN_RETRYTOKEN_EXPIRATION,
     MINIMUM_MAX_CRYPTO_BUFFER,
@@ -450,11 +475,123 @@ function validateQuicSocketOptions(options = {}) {
   };
 }
 
+function setConfigField(buffer, val, index) {
+  if (typeof val === 'number') {
+    buffer[index] = val;
+    return 1 << index;
+  }
+  return 0;
+}
+
+// Extracts configuration options and updates the aliased buffer
+// arrays that are used to communicate config choices to the c++
+// internals.
+function setTransportParams(config) {
+  const {
+    activeConnectionIdLimit,
+    maxStreamDataBidiLocal,
+    maxStreamDataBidiRemote,
+    maxStreamDataUni,
+    maxData,
+    maxStreamsBidi,
+    maxStreamsUni,
+    idleTimeout,
+    maxPacketSize,
+    maxAckDelay,
+    maxCryptoBuffer,
+    h3: {
+      qpackMaxTableCapacity,
+      qpackBlockedStreams,
+      maxHeaderListSize,
+      maxPushes,
+    },
+  } = { h3: {}, ...config };
+
+  // flags is a bitmap that is used to communicate whether or not a given
+  // configuration value has been explicitly provided.
+  const flags = setConfigField(sessionConfig,
+                               activeConnectionIdLimit,
+                               IDX_QUIC_SESSION_ACTIVE_CONNECTION_ID_LIMIT) |
+                setConfigField(sessionConfig,
+                               maxStreamDataBidiLocal,
+                               IDX_QUIC_SESSION_MAX_STREAM_DATA_BIDI_LOCAL) |
+                setConfigField(sessionConfig,
+                               maxStreamDataBidiRemote,
+                               IDX_QUIC_SESSION_MAX_STREAM_DATA_BIDI_REMOTE) |
+                setConfigField(sessionConfig,
+                               maxStreamDataUni,
+                               IDX_QUIC_SESSION_MAX_STREAM_DATA_UNI) |
+                setConfigField(sessionConfig,
+                               maxData,
+                               IDX_QUIC_SESSION_MAX_DATA) |
+                setConfigField(sessionConfig,
+                               maxStreamsBidi,
+                               IDX_QUIC_SESSION_MAX_STREAMS_BIDI) |
+                setConfigField(sessionConfig,
+                               maxStreamsUni,
+                               IDX_QUIC_SESSION_MAX_STREAMS_UNI) |
+                setConfigField(sessionConfig,
+                               idleTimeout * 1000000,  // millis to nano
+                               IDX_QUIC_SESSION_IDLE_TIMEOUT) |
+                setConfigField(sessionConfig,
+                               maxAckDelay,
+                               IDX_QUIC_SESSION_MAX_ACK_DELAY) |
+                setConfigField(sessionConfig,
+                               maxPacketSize,
+                               IDX_QUIC_SESSION_MAX_PACKET_SIZE) |
+                setConfigField(sessionConfig,
+                               maxCryptoBuffer,
+                               IDX_QUIC_SESSION_MAX_CRYPTO_BUFFER);
+
+  sessionConfig[IDX_QUIC_SESSION_CONFIG_COUNT] = flags;
+
+  const h3flags = setConfigField(http3Config,
+                                 qpackMaxTableCapacity,
+                                 IDX_HTTP3_QPACK_MAX_TABLE_CAPACITY) |
+                  setConfigField(http3Config,
+                                 qpackBlockedStreams,
+                                 IDX_HTTP3_QPACK_BLOCKED_STREAMS) |
+                  setConfigField(http3Config,
+                                 maxHeaderListSize,
+                                 IDX_HTTP3_MAX_HEADER_LIST_SIZE) |
+                  setConfigField(http3Config,
+                                 maxPushes,
+                                 IDX_HTTP3_MAX_PUSHES);
+
+  http3Config[IDX_HTTP3_CONFIG_COUNT] = h3flags;
+}
+
+// Some events that are emitted originate from the C++ internals and are
+// fairly expensive and optional. An aliased array buffer is used to
+// communicate that a handler has been added for the optional events
+// so that the C++ internals know there is an actual listener. The event
+// will not be emitted if there is no handler.
+function toggleListeners(self, event, on = 1) {
+  if (self[kHandle] === undefined || self.listenerCount(event) !== 0)
+    return;
+  switch (event) {
+    case 'keylog':
+      self[kHandle].state[IDX_QUIC_SESSION_STATE_KEYLOG_ENABLED] = on;
+      break;
+    case 'clientHello':
+      self[kHandle].state[IDX_QUIC_SESSION_STATE_CLIENT_HELLO_ENABLED] = on;
+      break;
+    case 'pathValidation':
+      self[kHandle].state[IDX_QUIC_SESSION_STATE_PATH_VALIDATED_ENABLED] = on;
+      break;
+    case 'OCSPRequest':
+      self[kHandle].state[IDX_QUIC_SESSION_STATE_CERT_ENABLED] = on;
+      break;
+  }
+}
+
 module.exports = {
   getAllowUnauthorized,
   getSocketType,
   lookup4,
   lookup6,
+  setTransportParams,
+  toggleListeners,
   validateCloseCode,
   validateNumberInRange,
   validateTransportParams,
